refactor : 폴더구조 변경에 따른 라우터 수정
