+++ src/pages/Posting/BoastDetail.tsx
