+++ src/pages/Posting/WritePost.tsx
@@ -83,8 +83,8 @@ const WritePost = () => {
       }
     }
     formData.append('tag', JSON.stringify(data.tags));
-    formData.append('content', data.content);
     formData.append('title', data.title);
+    formData.append('content', data.content.replace(/\n/g, '<br>'));
 
     formData.append('author_id', localStorage.getItem('authId') || '');
 
@@ -147,7 +147,7 @@ const WritePost = () => {
             placeholder="본문을 입력하세요"
             name="content"
             onChange={(e) => handleInputChange('content', e.target.value)}
-            className="mb-3 h-[25rem] rounded-xl border border-gray-200 p-5 font-[Pretendard] text-[0.875rem] font-normal not-italic leading-5 placeholder:text-gray-200"
+            className="mb-3 h-[25rem] whitespace-pre-wrap rounded-xl border border-gray-200 p-5 font-[Pretendard] text-[0.875rem] font-normal not-italic leading-5 placeholder:text-gray-200"
           ></textarea>
         </div>
         {error && (
